Moved Notification_profile to own app :truck: :construction:
