+++ src/aas/site/alert/models.py
@@ -1,5 +1,5 @@
 from django.db import models
-from ..auth.models import User
+
 
 class NetworkSystemType(models.Model):
     class Meta:
@@ -120,16 +120,3 @@ class Alert(models.Model):
 
     def __str__(self):
         return f"{self.timestamp}; {self.problem_type}: {self.object}"
-
-
-
-class Notification_profile(models.Model):
-    user = models.ForeignKey(
-        User,
-        models.CASCADE,
-    )
-    name = models.CharField(max_length=40)
-    interval_start = models.DateTimeField()
-    interval_stop = models.DateTimeField()
-
-
