+++ src/aas/site/notificationprofile/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from .models import Notification_profile
+
+
+admin.site.register(Notification_profile)
