+++ src/aas/site/notificationprofile/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class NotificationprofileConfig(AppConfig):
+    name = 'aas.site.notificationprofile'
