+++ src/aas/site/notificationprofile/models.py
@@ -1,0 +1,13 @@
+from django.db import models
+
+from ..auth.models import User
+
+
+class Notification_profile(models.Model):
+    user = models.ForeignKey(
+        User,
+        models.CASCADE,
+    )
+    name = models.CharField(max_length=40)
+    interval_start = models.DateTimeField()
+    interval_stop = models.DateTimeField()
