+++ src/aas/site/notificationprofile/tests.py
@@ -1,0 +1,1 @@
+# Create your tests here.
