+++ src/aas/site/settings/base.py
@@ -45,6 +45,7 @@ INSTALLED_APPS = [
 
     'aas.site.auth',
     'aas.site.alert',
+    'aas.site.notificationprofile',
 
     'rest_framework',
     'rest_framework.authtoken',
